requirements.txt
 
Flask
pymongo[srv]
python-dotenv
Werkzeug
PyJWT
openai